Build on Mac, but without fadvise
