+++ fileutil.go
@@ -13,7 +13,6 @@ import (
 	"os"
 	"runtime"
 	"sync"
-	"syscall"
 	"time"
 )
 
@@ -169,16 +168,3 @@ const (
 	POSIX_FADV_DONTNEED                        // Don't need these pages.  
 	POSIX_FADV_NOREUSE                         // Data will be accessed once.  
 )
-
-// Fadvise predeclares an access pattern for file data.
-// See also 'man 2 posix_fadvise'.
-func Fadvise(f *os.File, off, len int64, advice FadviseAdvice) (err error) {
-	_, _, errno := syscall.Syscall6(
-		syscall.SYS_FADVISE64,
-		uintptr(f.Fd()),
-		uintptr(off),
-		uintptr(len),
-		uintptr(advice),
-		0, 0)
-	return os.NewSyscallError("SYS_FADVISE64", errno)
-}
